Video Synopsis/Video1/requirements.txt
Requirement Libraries:

1-numpy == 1.18.1

2-cv2 == 3.4.2

3-matplotlib == 3.1.3

4-munkres == 1.1.2

5-collections == 
